+++ Astro/src/home/home.html
@@ -38,7 +38,7 @@
     <div class="sidebar">
       <div class="logo-details">
         <img
-          src="/Assets/Images/Icons/iconAstro.png"
+          src="../../Assets/Images/favicon.ico"
           class="logo_img"
           alt="astronaut"
         />
@@ -46,14 +46,14 @@
       </div>
       <ul class="nav-list">
         <li>
-          <a href="/Astro/Frontend/Html/Base/home.html">
+          <a href="./home.css">
             <i class="bx bx-home"></i>
             <span class="links_name">Home</span>
           </a>
           <span class="tooltip">Home</span>
         </li>
         <li>
-          <a href="/Astro/Frontend/Html/Others/search.html">
+          <a href="../search/search.html">
             <i class="bx bx-search"></i>
             <input type="text" placeholder="Search..." />
             <span class="tooltip">Search</span>
