+++ Astro/src/playlist/mixes/mixes.html
@@ -4,15 +4,11 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>ASTRO</title>
-    <link
-      rel="shortcut icon"
-      href="/Assets/Images/Icons/iconAstro.png"
-      type="image/x-icon"
-    />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/mainPages/mixes.css" />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/root.css" />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/root/player.css" />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/root/menus.css" />
+    <link rel="shortcut icon" href="../../../Assets/Images/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="../../../root/css/root.css">
+    <link rel="stylesheet" href="../../../root/css/player.css">
+    <link rel="stylesheet" href="../../../root/css/menus.css">
+    <link rel="stylesheet" href="./mixes.css">
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -140,12 +136,12 @@
     </section>
 
     <!--Backend srcipts-->
-    <script src="/Astro/Frontend/JS/others/gradientPlaylist.js"></script>
-    <script type="module" src="/Astro/Backend/JS/musicPlayer.js"></script>
-    <script type="module" src="/Astro/Backend/JS/Mixes/mixFuction.js"></script>
+    <script type="module" src="/Astro/Settings/libs/musicPlayer/musicPlayer.js"></script>
+    <script type="module" src="/Astro/Settings/libs/mixes/mixFuction.js"></script>
 
     <!--Frontend srcipts-->
-    <script src="/Astro/Frontend/JS/redirects/redirect.js"></script>
-    <script src="/Astro/Frontend/JS/others/menu.js"></script>
+    <script src="/Astro/root/js/menu.js"></script>
+    <script src="/Astro/root/js/gradientPlaylist.js"></script>
+    <script src="../../home/redirect.js"></script>
   </body>
 </html>
