+++ Astro/src/search/search.css
@@ -62,18 +62,7 @@ section.module.container.search label {
     outline: none;
 }
 
-/* Container Search Recents*/
-
-section.module.container.search-recent {
-
-}
-
-/* Container Sessons*/
-
-section.module.container.sessons {
-
-}
-
+x
 .container-results {
     margin: 20px 0;
 }
