+++ Astro/src/search/search.html
@@ -4,22 +4,11 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>ASTRO</title>
-    <link
-      rel="shortcut icon"
-      href="/Assets/Images/Icons/iconAstro.png"
-      type="image/x-icon"
-    />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/mainPages/search.css" />
-    <link rel="stylesheet" href="/Astro/Frontend/Css/root.css" />
-
-    <!--Player -->
-    <link
-      id="normal"
-      rel="stylesheet"
-      href="/Astro/Frontend/Css/root/player.css"
-    />
-
-    <link rel="stylesheet" href="/Astro/Frontend/Css/root/menus.css" />
+    <link rel="shortcut icon" href="../../Assets/Images/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="../../root/css/root.css">
+    <link rel="stylesheet" href="../../root/css/player.css">
+    <link rel="stylesheet" href="../../root/css/menus.css">
+    <link rel="stylesheet" href="./search.css">
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -129,13 +118,13 @@
       </main>
     </div>
     <footer></footer>
-    
+
     <!--Backend srcipts-->
-    <script type="module" src="/Astro/Backend/JS/musicPlayer.js"></script>
-    <script type="module" src="/Astro/Backend/JS/Settings/seach.js"></script>
+    <script type="module" src="/Astro/Settings/libs/musicPlayer/musicPlayer.js"></script>
+    <script type="module" src="/Astro/Settings/libs/search/seach.js"></script>
 
     <!--Frontend srcipts-->
-    <script src="/Astro/Frontend/JS/redirects/redirect.js"></script>
-    <script src="/Astro/Frontend/JS/others/menu.js"></script>
+    <script src="/Astro/root/js/menu.js"></script>
+    <script src="../../home/redirect.js"></script>
   </body>
 </html>
